🚚 change location commingsoon product
